# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Historial Clinico Backend",
        "description": "Aplicación backend del sistema web Historial Clnico",
        "contact": {
            "email": "dav.almeyda@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/historial-clinico-backend/public/api/obs/tratamiento": {
            "post": {
                "tags": [
                    "Observaciones"
                ],
                "summary": "Obtener la observación de un tratamiento",
                "operationId": "tratamiento",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal",
                                    "numAtendMed"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "numAtendMed": {
                                        "type": "string",
                                        "example": "0000384457"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/obs/procedimiento": {
            "post": {
                "tags": [
                    "Observaciones"
                ],
                "summary": "Obtener la observación de un procedimiento",
                "operationId": "procedimiento",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal",
                                    "numAtendMed"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "numAtendMed": {
                                        "type": "string",
                                        "example": "0000384457"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/obs/imagenes": {
            "post": {
                "tags": [
                    "Observaciones"
                ],
                "summary": "Obtener la observación de imagenes",
                "operationId": "imagenes",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal",
                                    "numAtendMed"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "numAtendMed": {
                                        "type": "string",
                                        "example": "0000384457"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
        "/historial-clinico-backend/public/api/obs/laboratorio": {
            "post": {
                "tags": [
                    "Observaciones"
                ],
                "summary": "Obtener la observación del laboratorio",
                "operationId": "laboratorio",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal",
                                    "numAtendMed"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "numAtendMed": {
                                        "type": "string",
                                        "example": "0000384457"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
...
        "/historial-clinico-backend/public/api/tratamientos": {
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener tratamientos",
                "operationId": "tratamientos",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codGrupoCia",
                                    "codLocal"
                                ],
                                "properties": {
                                    "codGrupoCia": {
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "codLocal": {
                                        "type": "string",
                                        "example": "001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos Encontrados"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}